refactor: create database module, rename databasetype to database service, add get all users function, add database module to user module
