+++ src/db/database.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { newDb } from './database.service';
+
+@Module({
+  providers: [newDb],
+  exports: [newDb],
+})
+export class DatabaseModule {}
