+++ src/users/user.service.ts
@@ -5,7 +5,7 @@ import { WrongPasswordException } from 'src/errors/wrongPassword';
 import { User, SafeUser } from 'src/users/userTypes';
 import { CreateUserDto, UpdateUserDto } from 'src/users/userDtos';
 import { validate, v4 as uuidv4 } from 'uuid';
-import { newDb } from 'src/db/databaseType';
+import { newDb } from 'src/db/database.service';
 
 @Injectable()
 export class UsersService {
@@ -13,6 +13,7 @@ export class UsersService {
 
   getAllUsers(): SafeUser[] {
     const users = this.db.getAllUsers();
+    console.log(users);
     const safeUsers = users.map((user) => {
       delete user.password;
       return user;
